# apps/desktop/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "COQ_GEN Core API",
    "summary": "Local sidecar: orchestration, DB access, COQ engine (Phase 0 scaffold).",
    "version": "0.1.0"
  },
  "paths": {
...
    "/batches": {
      "get": {
        "tags": [
          "batches"
        ],
        "summary": "List Batches",
        "operationId": "list_batches_batches_get",
        "parameters": [
          {
            "name": "x-coqgen-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Coqgen-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BatchSummary"
                  },
                  "title": "Response List Batches Batches Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/batches/{batch_no}": {
      "get": {
        "tags": [
          "batches"
        ],
        "summary": "Get Batch",
        "operationId": "get_batch_batches__batch_no__get",
        "parameters": [
          {
            "name": "batch_no",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch No"
            }
          },
          {
            "name": "x-coqgen-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Coqgen-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Batch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/batches/{batch_no}/master-parameters": {
      "get": {
        "tags": [
          "batches"
        ],
        "summary": "Master Parameters",
        "operationId": "master_parameters_batches__batch_no__master_parameters_get",
        "parameters": [
          {
            "name": "batch_no",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch No"
            }
          },
          {
            "name": "x-coqgen-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Coqgen-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MasterParameterLine"
                  },
                  "title": "Response Master Parameters Batches  Batch No  Master Parameters Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/register": {
      "get": {
        "tags": [
          "register"
        ],
        "summary": "Register",
        "operationId": "register_register_get",
        "parameters": [
          {
            "name": "x-coqgen-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Coqgen-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RegisterEntry"
                  },
                  "title": "Response Register Register Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/oos": {
      "get": {
        "tags": [
          "register"
        ],
        "summary": "Oos",
        "operationId": "oos_oos_get",
        "parameters": [
          {
            "name": "x-coqgen-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Coqgen-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OOSItem"
                  },
                  "title": "Response Oos Oos Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/labs": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Labs",
        "operationId": "labs_labs_get",
        "parameters": [
          {
            "name": "x-coqgen-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Coqgen-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LabInstitution"
                  },
                  "title": "Response Labs Labs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/parameters": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Parameters",
        "operationId": "parameters_parameters_get",
        "parameters": [
          {
            "name": "x-coqgen-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Coqgen-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ParameterDictionaryEntry"
                  },
                  "title": "Response Parameters Parameters Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Batch": {
        "properties": {
          "production_batch_number": {
            "type": "string",
            "title": "Production Batch Number"
          },
          "product_name": {
            "type": "string",
            "title": "Product Name"
          },
          "strain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strain"
          },
          "dominance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dominance"
          },
          "grade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade"
          },
          "grade_designation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade Designation"
          },
          "spec_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spec Code"
          },
          "spec_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spec Version"
          },
          "spec_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spec Reference"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "production_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Production Date"
          },
          "quantity_kg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity Kg"
          },
          "lineage": {
            "$ref": "#/components/schemas/LineageRef"
          }
        },
        "type": "object",
        "required": [
          "production_batch_number",
          "product_name",
          "status",
          "lineage"
        ],
        "title": "Batch"
      },
      "BatchSummary": {
        "properties": {
          "production_batch_number": {
            "type": "string",
            "title": "Production Batch Number"
          },
          "packaging_batch_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Packaging Batch Number"
          },
          "product_name": {
            "type": "string",
            "title": "Product Name"
          },
          "dominance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dominance"
          },
          "grade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade"
          },
          "grade_designation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade Designation"
          },
          "spec_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spec Reference"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "params_total": {
            "type": "integer",
            "title": "Params Total",
            "default": 0
          },
          "params_conforming": {
            "type": "integer",
            "title": "Params Conforming",
            "default": 0
          },
          "params_pending": {
            "type": "integer",
            "title": "Params Pending",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "production_batch_number",
          "product_name",
          "status"
        ],
        "title": "BatchSummary"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LabInstitution": {
        "properties": {
          "lab_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lab Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "credentials": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credentials"
          },
          "default_language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Language"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "LabInstitution"
      },
...
      "MasterParameterLine": {
        "properties": {
          "canonical_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Key"
          },
          "param_name": {
            "type": "string",
            "title": "Param Name"
          },
          "method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Method"
          },
          "acceptance_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acceptance Text"
          },
          "result_display": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Display"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "verdict": {
            "type": "string",
            "title": "Verdict",
            "default": "pending"
          },
          "source_institution_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Institution Name"
          },
          "source_lab_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Lab Code"
          },
          "source_document_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Document Code"
          },
          "source_document_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Document Date"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "draft"
          }
        },
        "type": "object",
        "required": [
          "param_name"
        ],
        "title": "MasterParameterLine",
        "description": "Single source-of-truth parameter row for the Batch Record + COQ lines."
      },
      "OOSItem": {
        "properties": {
          "batch_no": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch No"
          },
          "canonical_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Key"
          },
          "param_name": {
            "type": "string",
            "title": "Param Name"
          },
          "result_display": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Display"
          },
          "acceptance_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acceptance Text"
          },
          "source_document_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Document Code"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "open"
          }
        },
        "type": "object",
        "required": [
          "param_name"
        ],
        "title": "OOSItem",
        "description": "An out-of-specification result blocking CoQ issuance (OOS/NCR view)."
      },
      "ParameterDictionaryEntry": {
        "properties": {
          "canonical_key": {
            "type": "string",
            "title": "Canonical Key"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "canonical_unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Unit"
          },
          "default_method_family": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Method Family"
          }
        },
        "type": "object",
        "required": [
          "canonical_key",
          "display_name",
          "category"
        ],
        "title": "ParameterDictionaryEntry"
      },
...
      "RegisterEntry": {
        "properties": {
          "cert_type": {
            "type": "string",
            "title": "Cert Type"
          },
          "certificate_number": {
            "type": "string",
            "title": "Certificate Number"
          },
          "seq_no": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seq No"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          },
          "batch_no": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch No"
          },
          "product_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Name"
          },
          "spec_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spec Ref"
          },
          "issue_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issue Date"
          },
          "prepared_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prepared By"
          },
          "reviewed_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed By"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "oos_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oos Ref"
          },
          "archive_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archive Ref"
          }
        },
        "type": "object",
        "required": [
          "cert_type",
          "certificate_number",
          "status"
        ],
        "title": "RegisterEntry",
        "description": "Certificate Issuance Register row (iCoA/eCoA/CoQ) \u2014 QCSOP 012 v3."
      },
...
    }
  }
}